added option for full auto for future, cleaned up imports and got file running, FOUND BUG: if no input is entered the comparisison fails, ensure this is fixed
