+++ logic_file.py
@@ -1,6 +1,10 @@
 """File for the main meat of the application """
 from os import system, name
-from globals import EXIT_MESSAGE, EXIT_WORDS, ALGORITHM_OPTIONS
+from globals import (
+    EXIT_MESSAGE, EXIT_WORDS, ALGORITHM_OPTIONS, OPTIONS
+)
+from random import randint
+from time import sleep
 
 
 #helper functions
@@ -56,3 +60,20 @@ def check_exit(skip_line=True):
     if keywords.lower() in EXIT_WORDS:
         exit_function()
     return keywords
+
+def get_input_size():
+    # function that prompts for user input
+    print("How many entities would you like to use?")
+    print("Using default 2,000")
+    sleep(OPTIONS["CLEAR_SPEED"])
+    # return a list of viable options
+    optional_test_sizes = [2000, 10000, 50000, 100000]
+    # prompt user; tell all sizes and have choose numerically
+    # error check
+    # accept input and log to variable
+    clear()
+
+    unsorted_data = []
+    for _ in range(optional_test_sizes[0]):
+        unsorted_data.append(randint(1, 100000))
+    return unsorted_data
